d2::image, d2::image_rw, d2::ppm: Begin migration to libale types.  This breaks the build, for two reasons:
(1) When the build is broken, the compiler will instruct on areas where code
needs to be updated for the new libale object types.

(2) Revising code in a way allowing for existing code to function (with a
working build) alongside new code invites bugs, and preserves code not intended
for long-term use.

Working code can be found in current stable and testing versions, as well as in
the version just prior to this commit.

Libale, required for future development work in this tree, can be found at:

http://repo.or.cz/w/libale.git

Discussion of approaches for migration to use of libale and implementation
details can be found in the ditz issue tracker databases of ALE and Libale git
project trees.

More details about ditz can be found at

http://ditz.rubyforge.org/

(Packages for ditz are available, e.g., from Ubuntu repositories.)